refactor: extracted token refresh to new controller
